% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_knn.R
\name{calculateCosine}
\alias{calculateCosine}
\title{Cosine dissimilarity}
\usage{
calculateCosine(a, b)
}
\arguments{
\item{a:}{One of two vectors which are taken into consideration}

\item{b:}{Second of the vectors}
}
\description{
Calculates Cosine dissimilarity
}

