% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_knn.R
\name{knnCustom.closestNeighbours}
\alias{knnCustom.closestNeighbours}
\title{Neighbours generator for kNN}
\usage{
knnCustom.closestNeighbours(trainSet, testSet, simMeasure = 0, k = 1,
  isKplus = FALSE)
}
\arguments{
\item{simMeasure}{0: euclidean Distance, 1: cosine Dissimilarity. DEFAULTs to 0.}

\item{trainSet:}{Training set}

\item{testSet:}{Testing set}

\item{k:}{number of neighbours to obtains as voters for a class. DEFAULTs to 1.}

\item{isKplus:}{if TRUE, knn+ is used else: knn is used. DEFAULTs to FALSE}
}
\description{
Computes the closestNeighbours for the vorting
}

