% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_knn.R
\name{knnCustom.predict}
\alias{knnCustom.predict}
\title{kNN+ classifier}
\usage{
knnCustom.predict(trainSet, testSet, trainClasses, k = 1, simMeasure = 0,
  isKplus = FALSE, weight = NULL)
}
\arguments{
\item{:}{trainSet Training set}

\item{:}{testSet Testing set}

\item{:}{trainClasses Labels of the training set}

\item{:}{k number of neighbours to obtains as voters for a class. DEFAULTs to 1.}

\item{:}{simMeasure 0: euclidean Distance, 1: cosine Dissimilarity. DEFAULTs to 0.}

\item{:}{isKplus if TRUE, knn+ is used else: knn is used. DEFAULTs to FALSE}

\item{:}{weights constant weights for voting for a given class. DEFAULTs to NULL}
}
\description{
Predicts samples class using kNN classifier
}

