% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_custom.R
\name{myNaiveBayes}
\alias{myNaiveBayes}
\title{Naive Bayes classificator}
\usage{
myNaiveBayes(data, types, smoothing = 0)
}
\arguments{
\item{data}{data.frame of data to learn the classifier in form of DataTermMatrix with "Yes" or "No"}

\item{types}{list of classes/types of examples given in data}

\item{smoothing}{Lapplace-smoothing paraneter Default to 0}
}
\description{
This function returns model for prediction using Naive Bayes 
Classificator
}
\examples{
mod <- myNaiveBayes(learn_set, learn_types, 0.0001)
pred <- predict(mod, test_set)
}
\keyword{bayes}
\keyword{classification}
\keyword{naive}

