% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_knn.R
\name{orderKPlus}
\alias{orderKPlus}
\title{kNN+ ordered array of votes}
\usage{
orderKPlus(x, k)
}
\arguments{
\item{x}{Array to be ordered}

\item{k}{Parameter of minimal length of the output array of indices}
}
\description{
Calculates the ordered arrray of simple votes, if knn+ is used
}

