% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_custom.R
\name{predict.myNaiveBayes}
\alias{predict.myNaiveBayes}
\title{Classification using Naive Bayes}
\usage{
\method{predict}{myNaiveBayes}(model, newdata)
}
\arguments{
\item{model}{model of Naive Bayes returned by naiveBayes function}

\item{newdata}{data to be classified}
}
\description{
This function returns predictions of given data types
}
\examples{
mod <- myNaiveBayes(learn_set, learn_types, 0.0001)
pred <- predict(mod, test_set)
}
\keyword{bayes}
\keyword{classification}
\keyword{naive}

