% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tfidf.calculateAverageTF}
\alias{tfidf.calculateAverageTF}
\title{Calculates normalized TF for a given frequencyMatrix}
\usage{
tfidf.calculateAverageTF(frequencyMatrix)
}
\arguments{
\item{frequencyMatrix:}{Values of the DTM for a training set}
}
\description{
Creates a new TF Matrix from frequency matrix
}

