% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tfidf.calculateLogIDF}
\alias{tfidf.calculateLogIDF}
\title{Calculates IDF for a given frequencyMatrix}
\usage{
tfidf.calculateLogIDF(frequencyMatrix)
}
\arguments{
\item{frequencyMatrix:}{Values of the DTM for a training set}
}
\description{
Creates a new IDF Matrix from frequency matrix
}

