% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tfidf.generate}
\alias{tfidf.generate}
\title{TF-IDF funcion}
\usage{
tfidf.generate(frequencyMatrix, normalization = 0)
}
\arguments{
\item{frequencyMatrix:}{Values of the DTM for a training set}

\item{normalization:}{weighting type fot TF*IDF, 0: without normalization, 1: with normalization. DEFAULTs as 0}
}
\description{
Creates a new tfidf Matrix from frequency matrix
}

